refactor: separated types in feature folder
