+++ src/shared/types/brands.ts
@@ -1,0 +1,4 @@
+export type TBrand = {
+  id: string;
+  name: string;
+};
