+++ src/types/product.ts
@@ -82,37 +82,6 @@ export type TCartListItemDB = {
   salePrice: number | null;
 };
 
-export type TBrand = {
-  id: string;
-  name: string;
-};
-
-export type TFilters = {
-  stockStatus: "all" | "inStock" | "outStock";
-  priceMinMax: [number, number];
-  priceMinMaxLimitation: [number, number];
-  brands: TFilterBrands[];
-};
-export type TFilterBrands = {
-  id: string;
-  name: string;
-  isSelected: boolean;
-};
-
-export type TListItem = {
-  id: string;
-  name: string;
-  isAvailable: boolean;
-  specialFeatures: string[];
-  images: string[];
-  price: number;
-  salePrice: number | null;
-  brand: {
-    id: string;
-    name: string;
-  };
-};
-
 export type TSpecification = {
   groupName: string;
   specs: {
